separate DetermineNumberOfItems and SwitchKey
this introduces a small amount of duplicate code, but it makes the
code also more straight forward to understand and increases
the correlation with the XSD...
